refactor mudancas para incluir logica de autenticacao
