+++ src/app/components/layout/login/login.component.html
@@ -7,9 +7,9 @@
           <h3 class="card-title text-white fw-bold">Login</h3>
           <br>
           <form (submit)="logar()">
-            <input type="email" class="form-control no-border custom-input text-white" id="inputEmail" placeholder="Email" name="emailUsuario" [(ngModel)]="loginData.emailUsuario">                           
+            <input type="email" class="form-control no-border custom-input text-white" id="inputEmail" placeholder="Email" name="emailUsuario" [(ngModel)]="loginData.username">                           
             <br>                               
-            <input type="password" class="form-control no-border custom-input input2 text-white" id="inputSenha" placeholder="Senha" name="senhaUsuario" [(ngModel)]="loginData.senhaUsuario">              
+            <input type="password" class="form-control no-border custom-input input2 text-white" id="inputSenha" placeholder="Senha" name="senhaUsuario" [(ngModel)]="loginData.password">              
             <a href="#" style="color: var(--green-primary);">Esqueceu a senha?</a>
             <br>
             <br>
