+++ src/app/models/login.ts
@@ -2,6 +2,7 @@ export class Login {
   idUsuario!: number;
   emailUsuario: string;
   senhaUsuario: string;
+  role!: string;
 
   constructor(emailUsuario: string, senhaUsuario: string) {
     this.emailUsuario = emailUsuario;
