+++ src/app/services/usuario.service.ts
@@ -2,7 +2,6 @@ import { Injectable, inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { Usuario } from '../models/usuario';
-import { LoginService } from './login.service';
 
 @Injectable({
   providedIn: 'root'
